Run SDXL's autoencoder in float32, and get an image out
The last step of the pipeline now works. A prompt reaches pixels instead of
reaching a latent and an error message.

The autoencoder could not run in half and this was never a bug in the decoder:
its own config carries force_upcast, and the numbers say why. On the
reference's own four step latent the activations grow through the up blocks --
about 84 at the mid block, then 570, then 4046 -- and one convolution of the
last one passes 65504, which is as far as half goes. Everything after it was a
NaN. It is a range problem rather than a precision one, which is why the
kernels accumulating in float never helped: the value overflowed as it was
stored.

Only the autoencoder is widened. The U-Net and both text encoders stay in
half, where they belong and where they measure fine. `VarBuilder` grew
`with_float_type` so one file can be read at two precisions, and the exporter
already wrote the VAE's weights unnarrowed, so this is the file's own
precision rather than a widening of it. The decoder carries the type it was
built in and casts the latent on the way past, so the sampler hands over half
and nothing in between has to know.

Against the reference image the decode now drifts 1.0e-3, where the half
decoder was 2.1e-2 on the one latent it could take at all -- pure noise -- and
NaN on any real one. 768 by 768 at 25 steps takes 6.3 seconds on an RTX 5060
Ti.

`Sdxl::decode` no longer reads the whole image back to the host to check it
for NaNs, which was the report this replaces.

+++ waifu/tests/sdxl_vae.rs
@@ -32,12 +32,15 @@ fn models_dir() -> PathBuf {
     PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../models")
 }
 
+/// The decoder is built in float32, which is what SDXL's autoencoder needs and what the exporter
+/// writes its weights as: read at DType::Float16 the file would be narrowed on the way in and the
+/// last up block would overflow.
 fn weights() -> VarBuilder {
     let package = ZipFile::open(models_dir().join("sdxl-base.llmpkg")).unwrap();
     VarBuilder::from_reader(
         &mut package.open_entry("model.bin").unwrap(),
         Device::Cuda,
-        DType::Float16,
+        DType::Float,
     )
     .unwrap()
 }
